+++ GAME/SimonState.h
@@ -4,7 +4,7 @@ class SimonState
 {
 public:
 	SimonState() {};
-	virtual void Update(float dt) = 0;
+	virtual void Update(DWORD dt) = 0;
 	virtual void HandleKeyboard() = 0;
 	int StateName;
 };
