+++ GAME/SimonWalkingState.h
@@ -7,7 +7,7 @@ class SimonWalkingState:public SimonState
 {
 public:
 	SimonWalkingState();
-	void Update(float dt);
+	void Update(DWORD dt);
 	void HandleKeyboard();
 };
